+++ karma.conf.js
@@ -9,10 +9,12 @@ module.exports = function(karma) {
       //'vendor/external.js',
       'test/**/*.spec.js',
       { pattern: 'test/locales/**/*.json*', watched: true, included: false, served: true},
+      { pattern: 'test/languages/**/*.json*', watched: true, included: false, served: true},
     ],
 
     proxies: {
-      '/locales': 'http://localhost:9876/base/test/locales'
+      '/locales': 'http://localhost:9876/base/test/locales',
+      '/languages': 'http://localhost:9876/base/test/languages'
     },
 
     reporters: [ 'spec'/*, 'coverage' */],